requirements.txt
Flask==3.1.1
python-dotenv==1.1.0
youtube-transcript-api==1.2.2
requests==2.32.4